[ci] Improve CI for English Breakfast top level
This commit improves CI support for English Breakfast. We now build a
minimal amount of software (boot ROM + AES smoke test) in order to:
- Get actually functional FPGA bitstreams for the CW305.
- Run a basic smoke test on the Verilated system in CI.
