+++ test/systemtest/earlgrey/silicon_creator_config.py
