+++ test/systemtest/earlgrey/test_sim_verilator.py
@@ -4,179 +4,14 @@
 
 import logging
 import re
-import subprocess
-from pathlib import Path
 
 import pytest
 
-from .. import config, silicon_creator_config, roms_config, utils
+from . import config, silicon_creator_config, roms_config
+from .. import utils
+from .. import test_sim_verilator_opentitan as ot
 
 log = logging.getLogger(__name__)
-
-
-class VerilatorSimEarlgrey:
-    UART0_SPEED = 7200  # see device/lib/arch/device_sim_verilator.c
-
-    def __init__(self,
-                 sim_path: Path,
-                 rom_vmem_path: Path,
-                 otp_img_path: Path,
-                 work_dir: Path,
-                 boot_timeout: int = 120,
-                 test_timeout: int = 600):
-        """ A verilator simulation of the Earl Grey toplevel """
-        assert sim_path.is_file()
-        self._sim_path = sim_path
-
-        self.p_sim = None
-
-        assert rom_vmem_path.is_file()
-        assert otp_img_path.is_file()
-        self._rom_vmem_path = rom_vmem_path
-        self._otp_img_path = otp_img_path
-
-        assert work_dir.is_dir()
-        self._work_dir = work_dir
-
-        self.boot_timeout = boot_timeout
-        self.test_timeout = test_timeout
-
-        self._log = logging.getLogger(__name__)
-
-        self._uart0_log = None
-        self.uart0_log_path = None
-        self.spi0_log_path = None
-
-    def run(self, flash_elf=None, extra_sim_args=[]):
-        """
-        Run the simulation
-        """
-
-        self.uart0_log_path = self._work_dir / 'uart0.log'
-
-        cmd_sim = [
-            self._sim_path, '--meminit=rom,' + str(self._rom_vmem_path),
-            '--meminit=otp,' + str(self._otp_img_path),
-            '+UARTDPI_LOG_uart0=' + str(self.uart0_log_path)
-        ]
-        cmd_sim += extra_sim_args
-
-        if flash_elf is not None:
-            assert flash_elf.is_file()
-            cmd_sim.append('--meminit=flash,' + str(flash_elf))
-
-        self.p_sim = utils.Process(cmd_sim,
-                                   logdir=self._work_dir,
-                                   cwd=self._work_dir,
-                                   startup_done_expect='Simulation running',
-                                   startup_timeout=10)
-        self.p_sim.run()
-
-        self._log.info("Simulation running")
-
-        # Find paths to simulated I/O devices
-        # UART (through the simulated terminal)
-        self._uart0 = None
-        uart0_match = self.p_sim.find_in_output(
-            re.compile(r'UART: Created (/dev/pts/\d+) for uart0\.'),
-            timeout=1,
-            from_start=True)
-        assert uart0_match is not None
-        self.uart0_device_path = Path(uart0_match.group(1))
-        assert self.uart0_device_path.is_char_device()
-        self._log.debug("Found uart0 device file at {}".format(
-            str(self.uart0_device_path)))
-
-        # UART0 as logged directly to file by the uartdpi module.
-        assert self.uart0_log_path.is_file()
-        self._uart0_log = open(str(self.uart0_log_path), 'rb')
-
-        # SPI
-        spi0_match = self.p_sim.find_in_output(
-            re.compile(r'SPI: Created (/dev/pts/\d+) for spi0\.'),
-            timeout=1,
-            from_start=True)
-        assert spi0_match is not None
-        self.spi0_device_path = Path(spi0_match.group(1))
-        assert self.spi0_device_path.is_char_device()
-        self._log.debug("Found spi0 device file at {}".format(
-            str(self.spi0_device_path)))
-
-        self.spi0_log_path = self._work_dir / 'spi0.log'
-        assert self.spi0_log_path.is_file()
-
-        # GPIO
-        self.gpio0_fifo_write_path = self._work_dir / 'gpio0-write'
-        assert self.gpio0_fifo_write_path.is_fifo()
-
-        self.gpio0_fifo_read_path = self._work_dir / 'gpio0-read'
-        assert self.gpio0_fifo_read_path.is_fifo()
-
-        self._log.info("Simulation startup completed.")
-
-    def uart0(self):
-        if self._uart0 is None:
-            log_dir_path = self._work_dir / 'uart0'
-            log_dir_path.mkdir()
-            log.info("Opening UART on device {} ({} baud)".format(
-                str(self.uart0_device_path), self.UART0_SPEED))
-            self._uart0 = utils.LoggingSerial(
-                str(self.uart0_device_path),
-                self.UART0_SPEED,
-                timeout=1,
-                log_dir_path=log_dir_path,
-                default_filter_func=utils.filter_remove_device_sw_log_prefix)
-
-        return self._uart0
-
-    def terminate(self):
-        """ Gracefully terminate the simulation """
-        if self._uart0 is not None:
-            self._uart0.close()
-
-        if self._uart0_log is not None:
-            self._uart0_log.close()
-
-        self.p_sim.send_ctrl_c()
-
-        # Give the process some time to clean up
-        self.p_sim.proc.wait(timeout=5)
-        assert self.p_sim.proc.returncode == 0
-
-        try:
-            self.p_sim.terminate()
-        except ProcessLookupError:
-            # process is already dead
-            pass
-
-    def find_in_output(self,
-                       pattern,
-                       timeout,
-                       filter_func=None,
-                       from_start=False):
-        """ Find a pattern in STDOUT or STDERR of the Verilator simulation """
-        assert self.p_sim
-
-        return self.p_sim.find_in_output(pattern,
-                                         timeout,
-                                         from_start=from_start,
-                                         filter_func=filter_func)
-
-    def find_in_uart0(self,
-                      pattern,
-                      timeout,
-                      filter_func=utils.filter_remove_device_sw_log_prefix,
-                      from_start=False):
-        assert self._uart0_log
-
-        try:
-            return utils.find_in_files([self._uart0_log],
-                                       pattern,
-                                       timeout,
-                                       filter_func=filter_func,
-                                       from_start=from_start)
-        except subprocess.TimeoutExpired:
-            return None
 
 
 @pytest.fixture(params=roms_config.TEST_ROMS_SELFCHECKING,
@@ -294,26 +129,6 @@ def app_silicon_creator_selfchecking(request, bin_dir):
 # and the test never finishes.
 
 
-def assert_selfchecking_test_passes(sim):
-    assert sim.find_in_output(
-        re.compile(r"SW test transitioned to SwTestStatusInTest.$"),
-        timeout=sim.boot_timeout,
-        filter_func=utils.filter_remove_sw_test_status_log_prefix
-    ) is not None, "Start of test indication not found."
-
-    log.debug("Waiting for pass string from device test")
-
-    result_match = sim.find_in_output(
-        re.compile(r'^==== SW TEST (PASSED|FAILED) ====$'),
-        timeout=sim.test_timeout,
-        filter_func=utils.filter_remove_sw_test_status_log_prefix)
-    assert result_match is not None, "PASSED/FAILED indication not found in test output."
-
-    result_msg = result_match.group(1)
-    log.info("Test ended with {}".format(result_msg))
-    assert result_msg == 'PASSED'
-
-
 def test_roms_selfchecking(tmp_path, bin_dir, rom_selfchecking):
     """
     Run a self-checking ROM image on a Earl Grey Verilator simulation
@@ -328,11 +143,11 @@ def test_roms_selfchecking(tmp_path, bin_dir, rom_selfchecking):
     rom_vmem_path = rom_selfchecking[0]
     otp_img_path = bin_dir / "sw/device/otp_img/otp_img_sim_verilator.vmem"
 
-    sim = VerilatorSimEarlgrey(sim_path, rom_vmem_path, otp_img_path, tmp_path)
+    sim = ot.VerilatorSimOpenTitan(sim_path, rom_vmem_path, otp_img_path, tmp_path)
 
     sim.run(flash_elf=None, extra_sim_args=rom_selfchecking[1])
 
-    assert_selfchecking_test_passes(sim)
+    ot.assert_selfchecking_test_passes(sim)
 
     sim.terminate()
 
@@ -351,11 +166,11 @@ def test_apps_selfchecking(tmp_path, bin_dir, app_selfchecking):
     rom_vmem_path = bin_dir / "sw/device/boot_rom/boot_rom_sim_verilator.scr.39.vmem"
     otp_img_path = bin_dir / "sw/device/otp_img/otp_img_sim_verilator.vmem"
 
-    sim = VerilatorSimEarlgrey(sim_path, rom_vmem_path, otp_img_path, tmp_path)
+    sim = ot.VerilatorSimOpenTitan(sim_path, rom_vmem_path, otp_img_path, tmp_path)
 
     sim.run(app_selfchecking[0], extra_sim_args=app_selfchecking[1])
 
-    assert_selfchecking_test_passes(sim)
+    ot.assert_selfchecking_test_passes(sim)
 
     sim.terminate()
 
@@ -378,16 +193,16 @@ def test_apps_selfchecking_silicon_creator(tmp_path, bin_dir,
     otp_img_path = bin_dir / "sw/device/otp_img/otp_img_sim_verilator.vmem"
 
     # Use a longer timeout for boot due to the overhead of signature verification.
-    sim = VerilatorSimEarlgrey(sim_path,
-                               rom_vmem_path,
-                               otp_img_path,
-                               tmp_path,
-                               boot_timeout=55 * 60)
+    sim = ot.VerilatorSimOpenTitan(sim_path,
+                                   rom_vmem_path,
+                                   otp_img_path,
+                                   tmp_path,
+                                   boot_timeout=55 * 60)
 
     sim.run(app_silicon_creator_selfchecking[0],
             extra_sim_args=app_silicon_creator_selfchecking[1])
 
-    assert_selfchecking_test_passes(sim)
+    ot.assert_selfchecking_test_passes(sim)
 
     sim.terminate()
 
@@ -401,7 +216,7 @@ def test_spiflash(tmp_path, bin_dir):
     rom_vmem_path = bin_dir / "sw/device/boot_rom/boot_rom_sim_verilator.scr.39.vmem"
     otp_img_path = bin_dir / "sw/device/otp_img/otp_img_sim_verilator.vmem"
 
-    sim = VerilatorSimEarlgrey(sim_path, rom_vmem_path, otp_img_path, tmp_path)
+    sim = ot.VerilatorSimOpenTitan(sim_path, rom_vmem_path, otp_img_path, tmp_path)
     sim.run()
 
     log.debug("Waiting for simulation to be ready for SPI input")
@@ -414,7 +229,7 @@ def test_spiflash(tmp_path, bin_dir):
     utils.load_sw_over_spi(tmp_path, spiflash, app_bin,
                            ['--verilator', sim.spi0_device_path])
 
-    assert_selfchecking_test_passes(sim)
+    ot.assert_selfchecking_test_passes(sim)
 
     sim.terminate()
 
@@ -431,7 +246,7 @@ def test_openocd_basic_connectivity(tmp_path, bin_dir, topsrcdir, openocd):
     rom_vmem_path = bin_dir / "sw/device/boot_rom/boot_rom_sim_verilator.scr.39.vmem"
     otp_img_path = bin_dir / "sw/device/otp_img/otp_img_sim_verilator.vmem"
 
-    sim = VerilatorSimEarlgrey(sim_path, rom_vmem_path, otp_img_path, tmp_path)
+    sim = ot.VerilatorSimOpenTitan(sim_path, rom_vmem_path, otp_img_path, tmp_path)
     sim.run()
 
     # Wait a bit until the system has reached the bootrom and a first arbitrary
